+++ lib/widgets/settings/thumbnails/collection_actions_editor_page.dart
@@ -19,7 +19,9 @@ class CollectionActionEditorPage extends StatelessWidget {
         Tab(text: l10n.settingsCollectionQuickActionTabBrowsing),
         QuickActionEditorBody<EntrySetAction>(
           bannerText: context.l10n.settingsCollectionBrowsingQuickActionEditorBanner,
-          allAvailableActions: const [EntrySetActions.collectionEditorBrowsing],
+          allAvailableActions: const [
+            EntrySetActions.collectionEditorBrowsing,
+          ],
           actionIcon: (action) => action.getIcon(),
           actionText: (context, action) => action.getText(context),
           load: () => settings.collectionBrowsingQuickActions,